add dynamic metrics tests
Prior to this commit, no large tests existed for testing dynamic
metrics. There's some code in Publish() that replaces dynamic namespace
elements with tags, and this should also be covered.

This commit adds a few basic tests to influx_large_test.go, specifically
testing metrics that contain one and two dynamic elements. It also
increases verbosity so that the person / system running the test
captures the dynamic element name and value in the test output.