fix publishing for metrics with >1 dynamic elems
Prior to this commit, Publish() could only handle a single dynamic
element per namespace. If a second element was introduced, Publish()
would still use the original index, despite the array now being one
element smaller.

This commit now tracks which iteration of the loop we're on, beginning
with 0, and subtracts that from the original index to compensate:

    ns = append(ns[:j-i], ns[j-i+1:]...)